% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ComputeQ}
\alias{ComputeQ}
\title{Compute the Q values for each cell.}
\usage{
ComputeQ(
  sc_object,
  assay='RNA',
  k=6,
)
}
\arguments{
\item{sc_object}{A seurat object containing assays, reduction, meta data etc.}

\item{assay}{The default assay name used when creating the pseudo-Seurat object.}

\item{k}{The size of neighborhood. Default is 6.}
}
\value{
A vector of Q values of all cells.
}
\description{
Compute the Q values for each cell.
}
