% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCluster.R
\name{ConsensusRare}
\alias{ConsensusRare}
\title{Obtain more stable clustering result by applying FindRare on shuffled datasets.}
\usage{
ConsensusRare(
  sc_object,
  assay = "RNA",
  reduction = "pca",
  dims = 1:50,
  k.param = 20,
  k = 6,
  Q_cut = 0.6,
  ratio = 0.2,
  reps = 30,
  user_num =NULL
)
}
\arguments{
\item{sc_object}{A seurat object containing assays, reduction, meta data etc.}

\item{assay}{The Assay object of the input Seurat object for inferring rare cell types. Default is 'RNA'. User can adjust according to data modality.}

\item{reduction}{Reduction to use as input for shuffling and nearest neighbor searching}

\item{dims}{Dimensions of reduction to use as input}

\item{k.param}{Defines k for the k-nearest neighbor algorithm}

\item{k}{The size of neighborhood in computing Q values for rare cell detection. k should be between 5 and the total nearest neighbors (e.g., 20) obtained in processing step. The default value is 6.}

\item{Q_cut}{The default Q value for retaining rare clusters. Q should be between 0.5 and 1. The default value is 0.6.}

\item{ratio}{The threshold for merging clusters. When the proportion of a cluster's connections with its neighboring clusters among its total connections is larger than this threshold, merge the two cell groups. The default value is 0.2.}

\item{reps}{The maximum iterations for data shuffling. Default value is 30.}

\item{user_num}{User defined cluster number.}
}
\value{
Cluster assignments of cells including major and rare cell types.
}
\description{
ConsensusRare identifies more stable rare cell types.
}
\examples{
## Starts from counts
library(RareQ)
library(Seurat)

## Data prepocessing
sc_object <- CreateSeuratObject(count=counts, project = "sc_object", min.cells = 3)
sc_object$percent.mt <- PercentageFeatureSet(sc_object, pattern = "^MT-")
sc_object <- subset(sc_object, percent.mt<20)
sc_object <- NormalizeData(sc_object)
sc_object <- FindVariableFeatures(sc_object, nfeatures=2000)
sc_object <- ScaleData(sc_object, do.center = T)
sc_object <- RunPCA(sc_object, npcs=50)
sc_object <- FindNeighbors(object = sc_object,
                           k.param = 20,
                           compute.SNN = F,
                           prune.SNN = 0,
                           reduction = "pca",
                           dims = 1:50,
                           force.recalc = F, return.neighbor = T)
## Find robust rare cell types
robust.clusters <- ConsensusRare(sc_object,
                                 assay = 'RNA',
                                 reduction = 'pca',
                                 dims = 1:50,
                                 k.param = 20,
                                 k = 6,
                                 Q_cut = 0.6,
                                 ratio = 0.2,
                                 reps = 30)

}
