% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CreateObject}
\alias{CreateObject}
\title{Create a pseudo-Seurat object to accommodate emerging data modalities in rare cell identification.}
\usage{
CreateObject(
  embed.matrix,
  assay.name='RNA',
  dims = NULL
)
}
\arguments{
\item{embed.matrix}{Low-dimensional cell embeddings like PCA, SVD ....}

\item{assay.name}{The default assay name used when creating the pseudo-Seurat object.}

\item{dims}{Dimensions used in finding nearest neighbors. Such as 1:50, 2:30,...}

\item{k.param}{The number of neighbors to be identified.}
}
\value{
pseudo-Seurat object that can be used by FindRare to infer rare cell populations.
}
\description{
Use provided low-dimensional embeddings processed from any single cell data.
}
