% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ComputeQ}
\alias{ComputeQ}
\title{Compute the Q values for each cell.}
\usage{
CreateObject(
  embed.matrix,
  assay.name='RNA',
  dims = NULL
)
}
\arguments{
\item{embed.matrix}{Low-dimensional cell embeddings like PCA, SVD ....}

\item{assay.name}{The default assay name used when creating the pseudo-Seurat object.}

\item{dims}{Dimensions used in finding nearest neighbors. Such as 1:50, 2:30,...}
}
\value{
pseudo-Seurat object that can be used by FindRare to infer rare cell populations.
}
\description{
Compute the Q values for each cell for users to choose rare cell clusters.
}
