% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRare.R
\name{FindRare}
\alias{FindRare}
\title{Identify rare cell populations in single cell genomic data and cell-segmented spatial data.}
\usage{
FindRare(
  sc_object,
  assay = "RNA",
)
}
\arguments{
\item{sc_object}{A seurat object containing assays, reduction, meta data etc.}

\item{assay}{The Assay object of the input Seurat object for inferring rare cell types. Default is 'RNA'. User can adjust according to data modality.}
}
\value{
Cluster assignments of cells including major and rare cell types.
}
\description{
MetaQ identifies rare cell types via mutual connectivity (Q)-based network propagation.
}
\examples{
## Starts from counts
library(RareQ)
library(Seurat)

## Data prepocessing
sc_object <- CreateSeuratObject(count=counts, project = "sc_object", min.cells = 3)
sc_object$percent.mt <- PercentageFeatureSet(sc_object, pattern = "^MT-")
sc_object <- subset(sc_object, percent.mt<20)
sc_object <- NormalizeData(sc_object)
sc_object <- FindVariableFeatures(sc_object, nfeatures=2000)
sc_object <- ScaleData(sc_object, do.center = T)
sc_object <- RunPCA(sc_object, npcs=50)
sc_object <- FindNeighbors(object = sc_object,
                           k.param = 20,
                           compute.SNN = F,
                           prune.SNN = 0,
                           reduction = "pca",
                           dims = 1:50,
                           force.recalc = F, return.neighbor = T)
## Find rare cell types
clusters <- FindRare(sc_object)

}
